requirements.txt
Django>=4.1,<4.2
wagtail>=4.2,<4.3
django-recaptcha
wagtail-django-recaptcha